Added note cache

+++ MTD/Repository/WorkingWithUsers.cs
@@ -18,10 +18,8 @@ public class WorkingWithUsers: IUserRepository
         var user = await _context.Users.Select(s => new User {Id = s.Id, Email = s.Email, Nickname = s.Nickname}).FirstOrDefaultAsync(u => u.Email == email);
 
         if (user == null)
-        {
             throw new Exception("User is not found!");
-        }
-        
+
         return user;
     }
 
@@ -40,6 +38,9 @@ public class WorkingWithUsers: IUserRepository
     public async Task ModifyUser(User model)
     {
         var user = await _context.Users.FirstOrDefaultAsync(u => u.Id == model.Id);
+        
+        if (user == null)
+            throw new Exception("User is not found!");
 
         user.Email = model.Email;
         user.Password = model.Password;
@@ -51,6 +52,9 @@ public class WorkingWithUsers: IUserRepository
     public async Task ChangeEmail(int userId, string email)
     {
         var user = await _context.Users.FirstOrDefaultAsync(u => u.Id == userId);
+        
+        if (user == null)
+            throw new Exception("User is not found!");
 
         user.Email = email;
 
@@ -60,6 +64,9 @@ public class WorkingWithUsers: IUserRepository
     public async Task DeleteUser(int? id)
     {
         var user = await _context.Users.FirstOrDefaultAsync(u => u.Id == id);
+        
+        if (user == null)
+            throw new Exception("User is not found!");
 
         _context.Users.Remove(user);
